+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased changes
+
+### Internals
+
+- GatewayScanFilter: Treat `tunnelling` and `routing` attributes as OR when both are `True`.
+
 ## 0.18.9 HS-color 2021-07-26
 
 ### Devices